qing and python funcs work
