+++ .env
@@ -1,4 +1,4 @@
 PORT=3000
 NODE_ENV=development
 PORT_TEST=5000
-REDIS_URL=127.0.0.1:6379
+REDIS_URL=redis://:pd2db1123d511c70d30804d48fc934ec3b6a064e770ff84154b16af4212d1e0c6@ec2-18-210-137-167.compute-1.amazonaws.com:7879
