+++ Procfile
@@ -1,1 +1,2 @@
 web:npm start
+worker:node worker.js
