+++ server.js
@@ -13,53 +13,13 @@ let workQueue = Queue(REDIS_URL)
 
 let db = {}
 
-workQueue.process((job,done)=>{
-    const { walkerParams, czmlId } = job.data
-    const parseBody = (json) => {
-      console.log(json)
-      let stringifiedJSON = JSON.stringify(json)
-      let ret = ''
-      for(var i = 0;i < stringifiedJSON.length;i++){
-        if (stringifiedJSON[i] == '"'){
-          ret += "\\"
-        }
-        ret += stringifiedJSON[i]
-      }
-      return ret
-    }
-    const preprocessedParams = parseBody(walkerParams)
-
-    // change to run walker script
-    const command = `python ./SatLib/walker_script.py ${preprocessedParams}`
-    //done(null,{'stdout':'stdout here', 'stderr':'stderr here', 'error':'error here'})
-     
-    exec(command,(error,stdout,stderr)=>{
-        console.log(`Error:${error}`)
-        console.log(`Stdout:${stdout}`)
-        console.log(`Stderr:${stderr}`)
-        done(null,{
-            czmlId:czmlId,
-            'stdout':stdout,
-            'error':error,
-            'stderr':stderr
-        });
-    })
-    
-});
-
-workQueue.on('completed',(job,result)=>{
-
-    const { czmlId, czmlData } = result
-    // console.log(result);
-    db = {...db, [czmlId]:czmlData}
-});
 
 app.use(express.json())
 app.use(express.static('public'))
 
 app.use('/',homeRouters)
 
-app.post('/test-exec', (req,res,next)=>{
+app.post('/satellite', (req,res,next)=>{
     console.log('received a request')
     const walkerParams = req.body.walkerParams
     // 65 - 122, 0 - 9
@@ -71,7 +31,7 @@ app.post('/test-exec', (req,res,next)=>{
 app.get('/jobs/:id',(req,res,next) => {
     const{id} = req.params
     if(db.hasOwnProperty(id)){
-        res.send({finishedProcessing:true,czmlData:db[id]});
+        res.send({finishedProcessing:true,'czmlData':db[id]});
     }else{
         res.send({finishedProcessing:false})
     }
@@ -84,3 +44,5 @@ app.listen(PORT, (error)=>{
         console.log(`Listening on port ${PORT}`)
     }
 })
+
+module.exports = {db:db}
